+++ locals.tf
@@ -1,7 +1,8 @@
 locals {
   default_failover_locations = {
     default = {
-      location = var.location
+      location       = var.location
+      zone_redundant = var.zone_redundancy_enabled
     }
   }
 }